+++ honeycomb-benches/benches/core/cmap2/sewing.rs
@@ -15,7 +15,8 @@
 
 // ------ IMPORTS
 
-use honeycomb_core::{utils::GridBuilder, CMap2, FloatType};
+use honeycomb_benches::FloatType;
+use honeycomb_core::{utils::GridBuilder, CMap2};
 use iai_callgrind::{
     library_benchmark, library_benchmark_group, main, FlamegraphConfig, LibraryBenchmarkConfig,
 };