+++ honeycomb-core/src/lib.rs
@@ -14,7 +14,6 @@
 //! Optional features can be enabled when compiling this crate:
 //!
 //! - `utils` -- provides additionnal implementations for map generation, benchmarking & debugging
-//! - `single_precision` -- uses `f32` instead of `f64` for coordinates representation in tests
 
 // ------ CUSTOM LINTS
 
@@ -40,6 +39,8 @@ pub mod utils;
 
 // ------ RE-EXPORTS
 
+// --- PUBLIC API
+
 pub use attributes::{
     collections::{AttrCompactVec, AttrSparseVec},
     traits::{AttributeBind, AttributeUpdate},
@@ -53,7 +54,7 @@ pub use cells::{
     orbits::{Orbit2, OrbitPolicy},
 };
 pub use cmap2::CMap2;
-pub use common::{CMapError, CoordsFloat, DartIdentifier, FloatType, NULL_DART_ID};
+pub use common::{CMapError, CoordsFloat, DartIdentifier, NULL_DART_ID};
 pub use spatial_repr::{
     coords::{Coords2, CoordsError},
     vector::Vector2,