+++ honeycomb-examples/examples/memory_usage/compute.rs
@@ -1,8 +1,8 @@
-use honeycomb_core::{utils::GridBuilder, CMap2, DartIdentifier, FloatType};
+use honeycomb_core::{utils::GridBuilder, CMap2, DartIdentifier};
 
 pub fn main() {
     // create a 3x3 grid & remove the central square
-    let mut cmap: CMap2<FloatType> = GridBuilder::unit_squares(3).build2().unwrap();
+    let mut cmap: CMap2<f64> = GridBuilder::unit_squares(3).build2().unwrap();
     // darts making up the central square
     let (d1, d2, d3, d4): (
         DartIdentifier,